pyproject.toml
[project]
name = "backgroundremover"
version = "0.1.0"
description = ""
authors = [
    {name = "Serhii Hrekov",email = "themebeat@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "rembg (>=2.0.67,<3.0.0)",
    "onnxruntime (>=1.22.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
